refactor(schema): Move PartialVersion into schemas
