+++ src/cargo/lib.rs
@@ -153,7 +153,6 @@ pub mod ops;
 pub mod sources;
 pub mod util;
 pub mod util_schemas;
-pub mod util_semver;
 mod version;
 
 pub fn exit_with_error(err: CliError, shell: &mut Shell) -> ! {
