+++ src/cargo/util_schemas/core/mod.rs
@@ -1,6 +1,8 @@
 mod package_id_spec;
+mod partial_version;
 mod source_kind;
 
 pub use package_id_spec::PackageIdSpec;
+pub use partial_version::PartialVersion;
 pub use source_kind::GitReference;
 pub use source_kind::SourceKind;
